requirements.txt
Flask
pymongo[srv]
dnspython
python-dotenv
Werkzeug
pytest
coverage
